ref(UserFeedback): Move captureUserFeedback fully in to the core integration
